refactor: Rebuild project with vanilla HTML, JS, and CSS
This commit refactors the entire project from a React/Vite application to a simple, self-contained application using vanilla HTML, JavaScript, and CSS.

This change was made to simplify the project structure and resolve persistent deployment issues related to the Node.js build environment.

- The React/Vite project and all its dependencies have been removed.
- A new `index.html` file has been created, which loads Leaflet.js from a CDN.
- All application logic has been moved into a new `main.js` file.
- All styles have been moved into a new `style.css` file.
- The interactive map and the proof-of-concept quest layer have been re-implemented using vanilla JavaScript and the Leaflet API.

The resulting application is now free of any build dependencies and can be directly served by GitHub Pages.
