--- src/App.css
@@ -1,5 +1,0 @@
-html, body, #root, .App {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-}
