--- src/App.jsx
@@ -1,13 +1,0 @@
-import React from 'react';
-import Map from './components/Map';
-import './App.css';
-
-function App() {
-  return (
-    <div className="App">
-      <Map />
-    </div>
-  );
-}
-
-export default App;
