--- src/index.css
@@ -1,1 +1,0 @@
-/* Reset and base styles will go here */
