--- vite.config.js
@@ -1,8 +1,0 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
-
-// https://vite.dev/config/
-export default defineConfig({
-  base: '/The-Loremasters-Compass/',
-  plugins: [react()],
-})
